server/requirements.txt
kafka-python
pymongo
scikit-learn
prometheus_client